Introduce Helper module for class methods
This abstracts the knowledge of looking up from Store
or falling back into a single place so Orchestrate and Replay
can both perform the same logic
